Separate main code in files
